requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-dateutil
pytz

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth-requests
requests
python-dateutil
dateparser
SQLAlchemy
aiosqlite



